refactor cython
