+++ .gitignore
@@ -7,7 +7,7 @@ __pycache__/
 *.so
 
 # Cython-generated C source files
-oasislmf/**/_write_csv_cython.c
+oasislmf/**/oasis_writecsv.c
 
 # Ktools binary files
 *.bin
